commenting configs and cleaning-up README

+++ Rakefile
@@ -7,28 +7,32 @@ IGNORE_FILES = [/^\.gitignore$/, /^Rakefile$/,/^README.textile$/]
 files = `git ls-files`.split("\n")
 files.reject! { |f| IGNORE_FILES.any? { |re| f.match(re) } }
 
-desc 'Installs dotfiles'
+desc 'installs dotfiles in home dir'
 task :install do
-  targetdir=File.expand_path("~")
-  files.each do |file|
-    target_file = File.join(targetdir, ".#{file}")
-    FileUtils.mkdir_p File.dirname(target_file)
-    FileUtils.cp file, target_file
+    targetdir=File.expand_path("~")
+    files.each do |file|
+        if File.exists?(file)
+            target_file = File.join(targetdir, ".#{file}")
+            FileUtils.mkdir_p File.dirname(target_file)
+            FileUtils.cp file, target_file
 
-    puts "Installed #{file} to #{target_file}"
-  end
+            puts "Installed #{file} to #{target_file}"
+        else
+            puts "#{file} removed?"
+        end
+    end
 
 end
 
-desc 'Pulls from git repository'
+desc 'pulls from git repository'
 task :pull do
-  puts "Updating from git repository"
-  system("cd " << Dir.new(File.dirname(__FILE__)).path << " && git pull")
+    puts "Updating from git repository"
+    system("cd " << Dir.new(File.dirname(__FILE__)).path << " && git pull")
 end
 
-desc 'Updates from git repository and then updates files in dir'
+desc 'updates from git repository and then updates files in dir'
 task :update => ['pull', 'install'] do
-  puts "Update of dotfiles completed."
+    puts "Update of dotfiles completed."
 end
 
 task :default => ['install']
